requirements.txt
pandas
numpy
openai
hdbscan
umap-learn
plotly
streamlit
scikit-learn
python-dotenv
tqdm
